split out bot commands
