+++ info_bot.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python -u
+#!/usr/bin/python
 import sys
 
 if __name__ == '__main__':
